+++ CHANGELOG.md
@@ -7,13 +7,14 @@
   | `{{'a' 10, 'b'}}`            | `{{a} 10, {b}}`              |
   | `{{use anotherMode}}`        | `{use anotherMode}`          |
   | `'foo' as 'bar' 5, call foo` | `{foo} as {bar} 5, call foo` |
+* Remove the `begin` statement; instead the first non-prelude-y text will cause the prelude to end. To start with an active mode, simply call the `{use someMode}` command.
+* Remove the `using` variant of the `use` keyword
 * Support recursive rendering within replacements, both in inline choices and in rule replacements. For instance:
   ```bml
   mode main {
       {recurse!} as {just kidding}, {outer {{inner 1}, {inner 2}}}
   }
-  begin using main
-
+  {use main}
   recurse!
   a {{simple inline}, {{{complex}, {fancy}} recursive}} inline choice
   ```